// frontend/app/(manga_narrator)/manga_narrator/types/tts_api_types.ts
export type Emotion =
    components["schemas"]["Emotion"]
export type EmotionParams =
    components["schemas"]["EmotionParams"]
export type Gender =
    components["schemas"]["Gender"]

export const GENDER_OPTIONS = [
    "female",
    "male",
    "neutral",
] as const

// export type MediaRef =
//     components["schemas"]["MediaRef"]
export type Speaker =
    components["schemas"]["Speaker"]
export type TTSInput =
    components["schemas"]["TTSInput-Input"]
export type TTSOutput =
    components["schemas"]["TTSOutput"]
export type EmotionOptionsOutput =
    components["schemas"]["EmotionOptionsOutput"]
// ...
export interface components {
    schemas: {
        /** Emotion */
        Emotion: {
            /** Name */
            name: string;
            params: components["schemas"]["EmotionParams"];
        };
        /** EmotionOptionsOutput */
        EmotionOptionsOutput: {
            /** Emotionoptions */
            emotionOptions: components["schemas"]["Emotion"][];
        };
        /** EmotionParams */
        EmotionParams: {
            /** Cfg */
            cfg: number;
            /** Exaggeration */
            exaggeration: number;
        };
        /** Gender */
        Gender: {
            /**
             * Value
             * @enum {string}
             */
            value: "female" | "male" | "neutral";
        };
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components["schemas"]["ValidationError"][];
        };
        /**
         * MediaNamespace
         * @enum {string}
         */
        MediaNamespace: "inputs" | "outputs";
        /** MediaRef */
        MediaRef: {
            namespace: components["schemas"]["MediaNamespace"];
            /** Path */
            path: string;
        };
        /** Speaker */
        Speaker: {
            /** Name */
            name: string;
            /** Wav File */
            wav_file: string;
            gender: components["schemas"]["Gender"];
        };
        /** TTSInput */
        "TTSInput-Input": {
            /** Text */
            text: string;
            gender: components["schemas"]["Gender"];
            emotion: components["schemas"]["Emotion"];
            speaker: components["schemas"]["Speaker"];
            image_ref: components["schemas"]["MediaRef"];
            customSettings?: components["schemas"]["EmotionParams"] | null;
            /**
             * Run Id
             * @default
             */
            run_id: string;
            /**
             * Custom Filename
             * @default
             */
            custom_filename: string;
            /**
             * Dialogue Id
             * @default -1
             */
            dialogue_id: number;
        };
        /** TTSInput */
        "TTSInput-Output": {
            /** Text */
            text: string;
            gender: components["schemas"]["Gender"];
            emotion: components["schemas"]["Emotion"];
            speaker: components["schemas"]["Speaker"];
            image_ref: components["schemas"]["MediaRef"];
            customSettings?: components["schemas"]["EmotionParams"] | null;
            /**
             * Run Id
             * @default
             */
            run_id: string;
            /**
             * Custom Filename
             * @default
             */
            custom_filename: string;
            /**
             * Dialogue Id
             * @default -1
             */
            dialogue_id: number;
        };
        /** TTSOutput */
        TTSOutput: {
            ttsInput: components["schemas"]["TTSInput-Output"];
            audio_ref: components["schemas"]["MediaRef"];
        };
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[];
            /** Message */
            msg: string;
            /** Error Type */
            type: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}